% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_score.R
\name{brier_score}
\alias{brier_score}
\title{Brier score}
\usage{
brier_score(pred_prob, binary_obs)
}
\arguments{
\item{pred_prob}{Predicted probabilities}

\item{binary_obs}{Binary observations}
}
\value{
Brier score
}
\description{
Brier score
}
