% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brierscore.R
\name{brierscore}
\alias{brierscore}
\title{Brier score}
\usage{
brierscore(pred_prob, binary_obs)
}
\arguments{
\item{pred_prob}{Predicted probabilities}

\item{binary_obs}{Binary observations}
}
\value{
Brier score
}
\description{
Brier score
}
