% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_change.R
\name{check_for_change}
\alias{check_for_change}
\title{Check for ramps/shutdowns within a univariate time series subset}
\usage{
check_for_change(data, ch)
}
\arguments{
\item{data}{A vector of numerical values}

\item{ch}{Flat signed change in forecasted variable}
}
\value{
Boolean, 1 = event happened, 0 = event did not happen
}
\description{
Check for ramps/shutdowns within a univariate time series subset
}
