% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForEvent.R
\name{checkForEvent}
\alias{checkForEvent}
\title{Check for ramps/shutdowns within a univariate time series subset}
\usage{
checkForEvent(data, ch)
}
\arguments{
\item{data}{Data object returned by the loadData function}

\item{ch}{Flat signed change in forecasted variable}
}
\value{
Boolean, 1 = event happened, 0 = event did not happen
}
\description{
Check for ramps/shutdowns within a univariate time series subset
}
