% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table.R
\name{contingency_table}
\alias{contingency_table}
\title{Contingency table}
\usage{
contingency_table(detect_table_binary)
}
\arguments{
\item{detect_table_binary}{A detection table of binary event forecasts (1 = event, 0 = no event).}
}
\value{
A contingency table for a forecast candidate.
}
\description{
Contingency table
}
