% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_list.R
\name{contingency_table_list}
\alias{contingency_table_list}
\title{List of contingency tables}
\usage{
contingency_table_list(detect_table_list, threshold = 0.75)
}
\arguments{
\item{detect_table_list}{A list of event detection tables.}

\item{threshold}{Required number of positive ensemble members. Defaults to 0.75.}
}
\value{
Contingency tables for all forecast candidates
}
\description{
List of contingency tables
}
