% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencyTable.R
\name{contingencyTable}
\alias{contingencyTable}
\title{Contingency table}
\usage{
contingencyTable(tab)
}
\arguments{
\item{tab}{TBD}
}
\value{
Contingency table for one forecast candidate
}
\description{
Contingency table
}
