% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencyTableList.R
\name{contingencyTableList}
\alias{contingencyTableList}
\title{Contingency table list}
\usage{
contingencyTableList(detect_table_list, threshold = 0.75)
}
\arguments{
\item{detect_table_list}{List of detection tables}

\item{threshold}{Required number of positive ensemble members}
}
\value{
Contingency table for all forecast candidates
}
\description{
Contingency table list
}
