% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_detect_to_binary.R
\name{convert_detect_to_binary}
\alias{convert_detect_to_binary}
\title{Convert continuous classifier into binary classifier w.r.t a threshold}
\usage{
convert_detect_to_binary(tab, threshold)
}
\arguments{
\item{tab}{Event detection table returned by the detect_event function.}

\item{threshold}{A continuous number between 0 and 1.}
}
\value{
Table of overall detection by a forecast candidate.
}
\description{
Convert continuous classifier into binary classifier w.r.t a threshold
}
