% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps.R
\name{crps}
\alias{crps}
\title{CRPS (univariate) of a probabilistic forecast w.r.t an observation.}
\usage{
crps(f, y)
}
\arguments{
\item{f}{Vector representing a probabilistic forecast.}

\item{y}{The corresponding observation.}
}
\value{
Continuous Ranked Probability Score (CRPS) of f and y.
}
\description{
CRPS (univariate) of a probabilistic forecast w.r.t an observation.
}
