% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_event.R
\name{detect_event}
\alias{detect_event}
\title{Function for detecting all events in a multivariate time series}
\usage{
detect_event(data, change = NA, range = NA, window)
}
\arguments{
\item{data}{Data object returned by the loadData function}

\item{change}{Flat signed change in forecasted variable}

\item{range}{?}

\item{window}{Time window length within which the change occurs}
}
\value{
Returns an event detection table
}
\description{
Function for detecting all events in a multivariate time series
}
