% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectToBinary.R
\name{detectToBinary}
\alias{detectToBinary}
\title{Convert continuous classifier into binary classifier w.r.t a threshold}
\usage{
detectToBinary(tab, threshold)
}
\arguments{
\item{tab}{Event detection table returned by the eventDetect function.}

\item{threshold}{A number.}
}
\value{
Table of overall detection by a forecast candidate.
}
\description{
Convert continuous classifier into binary classifier w.r.t a threshold
}
