% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_joint_distribution.R
\name{evaluate_joint_distribution}
\alias{evaluate_joint_distribution}
\title{Forecast evaluation of the joint distribution using the variogram score}
\usage{
evaluate_joint_distribution(data, by_base_time = T, p = 0.5)
}
\arguments{
\item{data}{A list containing forecasts and observations.}

\item{by_base_time}{Logical indicator for whether forecast evaluation should be base time-specific or not. Defaults to true.}
}
\value{
A table of scores for the dataset.
}
\description{
Forecast evaluation of the joint distribution using the variogram score
}
