% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_marginal_distribution.R
\name{evaluate_marginal_distribution}
\alias{evaluate_marginal_distribution}
\title{Forecast evaluation}
\usage{
evaluate_marginal_distribution(data, by_lead_time = T)
}
\arguments{
\item{data}{A list containing forecasts and observations.}

\item{by_lead_time}{Logical indicator for whether forecast evaluation should be lead time-specific or not. Defaults to true.}
}
\value{
A table of scores for the dataset.
}
\description{
Forecast evaluation
}
