% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluationSet.R
\name{evaluationSet}
\alias{evaluationSet}
\title{Refine data to evaluatable dataset}
\usage{
evaluationSet(data)
}
\arguments{
\item{data}{Dataset.}
}
\value{
A number.
}
\description{
Refine data to evaluatable dataset
}
