% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_detection_table.R
\name{event_detection_table}
\alias{event_detection_table}
\title{Compute a list of event detection tables from a list of forecast candidates}
\usage{
event_detection_table(
  data,
  change = NA,
  range = NA,
  window,
  export.results = F
)
}
\arguments{
\item{data}{Data object returned by the loadData function.}

\item{change}{Flat signed change in forecasted variable.}

\item{range}{Continuous range in which events occur.}

\item{window}{Duration of time window within which the change occurs, given in hours.}
}
\value{
Returns an event detection table.
}
\description{
Compute a list of event detection tables from a list of forecast candidates
}
