% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventDetect.R
\name{eventDetect}
\alias{eventDetect}
\title{Function for detecting all events in a multivariate time series}
\usage{
eventDetect(data, ch, win)
}
\arguments{
\item{data}{Data object returned by the loadData function}

\item{ch}{Flat signed change in forecasted variable}

\item{win}{Time window length within which the change occurs}
}
\value{
Returns an event detection table
}
\description{
Function for detecting all events in a multivariate time series
}
