% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventDetectionTable.R
\name{eventDetectionTable}
\alias{eventDetectionTable}
\title{Function for constructing list of event detection tables.}
\usage{
eventDetectionTable(data, change = NA, range = NA, window, export.results = F)
}
\arguments{
\item{data}{Data object returned by the loadData function.}

\item{change}{Flat signed change in forecasted variable.}

\item{range}{?}

\item{window}{Duration of time window within which the change occurs, given in hours.}
}
\value{
Returns an event detection table.
}
\description{
Function for constructing list of event detection tables.
}
