% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastEvaluation.R
\name{forecastEvaluation}
\alias{forecastEvaluation}
\title{Forecast evaluation}
\usage{
forecastEvaluation(data)
}
\arguments{
\item{data}{A list containing forecasts and observations.}
}
\value{
A table of scores for the dataset.
}
\description{
Forecast evaluation
}
