% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_forecast_evaluation_report.R
\name{generate_forecast_evaluation_report}
\alias{generate_forecast_evaluation_report}
\title{Report generation function}
\usage{
generate_forecast_evaluation_report(
  data,
  dest = "./",
  delete_source = T,
  event_change = 0.1,
  event_window = 1,
  contingency_threshold = 0.2
)
}
\arguments{
\item{data}{Data set in the required format.}

\item{dest}{Destination for the compiled HTML report.}
}
\value{
Currently nothing.
}
\description{
Report generation function
}
