% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lead_time.R
\name{get_lead_time}
\alias{get_lead_time}
\title{Extract lead time from forecast data.}
\usage{
get_lead_time(forecast)
}
\arguments{
\item{forecast}{A forecast table or row containing at least the columns named TimeStamp and BaseTime, respectively.}
}
\value{
Lead times of the input forecasts.
}
\description{
Extract lead time from forecast data.
}
