% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLeadTime.R
\name{getLeadTime}
\alias{getLeadTime}
\title{Add together two numbers}
\usage{
getLeadTime(forecast_row)
}
\arguments{
\item{forecast_row}{A row from a forecast table containing at least TimeStamp and BaseTime.}
}
\value{
Lead time of a forecast row.
}
\description{
Add together two numbers
}
