% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IEAW51-MSEPS-DKsynop48}
\alias{IEAW51-MSEPS-DKsynop48}
\alias{mseps_dk_thoroe48}
\title{Sample data included in package}
\format{
list with two entries, \code{observations} and \code{forecasts}. 
\code{forecasts} is a list of \code{data.frame}s containing forecast data.
}
\usage{
mseps_dk_thoroe48
}
\description{
Wind speed ensemble forecasts from the MSEPS Ensemble with 75 members/forecasts
and observations from the corresponding Synop Station in Denmark called Thorø.
}
\keyword{data}
