% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDetectionTable.R
\name{importDetectionTable}
\alias{importDetectionTable}
\title{Import detection tables.}
\usage{
importDetectionTable(path, tz = "UTC")
}
\arguments{
\item{path}{the path to the directory where the files are to be imported from. 
All files will be imported, so the directory must contain only exactly the files that are to be imported, nothing more, nothing less.}

\item{tz}{the timezone of the TimeStamp column in the files.}
}
\value{
a list of all the imported tables.
}
\description{
Import detection tables.
}
