% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecast_data.R
\name{load_forecast_data}
\alias{load_forecast_data}
\title{Load all forecast and observation data}
\usage{
load_forecast_data(
  path,
  timeformat_f = "default",
  timeformat_obs = "default",
  timezone_f = "UTC",
  timezone_obs = "UTC"
)
}
\arguments{
\item{dir}{A directory containing forecast and observation data.}
}
\value{
A list with two elements. The first element contains forecast data as a list, and the second contains observations.
}
\description{
Load all forecast and observation data
}
