% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load all forecast and observation data}
\usage{
loadData(dir)
}
\arguments{
\item{dir}{A directory containing forecast and observation data.}
}
\value{
A list with two elements. The first element contains forecast data as a list, and the second contains observations.
}
\description{
Load all forecast and observation data
}
