% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_score.R
\name{log_score}
\alias{log_score}
\title{LogS (univariate) of a probabilistic forecast w.r.t an observation.}
\usage{
log_score(f, y)
}
\arguments{
\item{f}{Vector representing a probabilistic forecast.}

\item{y}{The corresponding observation.}
}
\value{
Logarithmic score (LogS) of f and y.
}
\description{
LogS (univariate) of a probabilistic forecast w.r.t an observation.
}
