% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{logs}
\alias{logs}
\title{LogS (univariate) of a probabilistic forecast w.r.t an observation.}
\usage{
logs(f, y)
}
\arguments{
\item{y}{The corresponding observation.}

\item{sim}{Vector representing a probabilistic forecast.}
}
\value{
Logarithmic score (LogS) of f and y.
}
\description{
LogS (univariate) of a probabilistic forecast w.r.t an observation.
}
