% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_evaluation_subset.R
\name{make_evaluation_subset}
\alias{make_evaluation_subset}
\title{Refine data to evaluatable dataset}
\usage{
make_evaluation_subset(data, lead_time = 0)
}
\arguments{
\item{data}{A forecast-observation data structure.}

\item{lead_time}{The lead time of interest in units of hours. Defaults to 0, which corresponds to no lead time specified.}
}
\value{
A forecast-observation data structure with only intersecting timestamps preserved.
}
\description{
Refine data to evaluatable dataset
}
