% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_observations.R
\name{plot_observations}
\alias{plot_observations}
\title{Plot observations}
\usage{
plot_observations(
  data,
  x = NULL,
  grid = F,
  col = "black",
  all = F,
  numobs = 100,
  main = "Observations",
  xlab = "Time",
  ylab = "Quantity"
)
}
\arguments{
\item{data}{a data frame of univariate observations.}

\item{x}{Optional time stamps corresponding to rows of \code{data}.}

\item{grid}{Add a grid to the plot background?}

\item{col}{Color of the observations.}

\item{all}{Logical, if all observations should be drawn. Defaults to false, in which case up to 100 observations are shown.}

\item{main}{a main title for the plot, see also \code{?plot}.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{...}{Additional arguments passed to \code{plot()}}
}
\value{
Does not return anything, but displays a quantile plot.
}
\description{
Plot observations
}
