% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_by_leadtime.R
\name{plot_score_by_leadtime}
\alias{plot_score_by_leadtime}
\title{Plot score by lead time}
\usage{
plot_score_by_leadtime(
  scoretable,
  main = "default",
  xlab = "Lead time",
  ylab = "default",
  xlim = "default",
  ylim = "default",
  legend = T,
  legend_pos = "topleft"
)
}
\arguments{
\item{scoretable}{a table of scores returned from the forecastEvaluation function}

\item{main}{the title of the plot.}

\item{xlab}{the x-label.}

\item{ylab}{the y-label.}

\item{xlim}{the x-range. Selected automatically unless specified.}

\item{ylim}{the y-range. Selected automatically unless specified.}
}
\value{
a plot of scores by leadtime with one curve for each forecast candidate.
}
\description{
Plot score by lead time
}
