% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variogram_scores.R
\name{plot_variogram_scores}
\alias{plot_variogram_scores}
\title{Plot comparable variogram scores}
\usage{
plot_variogram_scores(
  scoretable,
  main = "Variogram score by basetime",
  xlab = "Basetime",
  ylab = "VarS",
  xlim = "default",
  ylim = "default",
  legend = T,
  legend_pos = "topleft"
)
}
\arguments{
\item{scoretable}{a table of scores returned from the forecastEvaluation function}

\item{main}{the title of the plot. Defaults to "Score by leadtime".}

\item{xlab}{the x-label. Defaults to "Leadtime".}

\item{ylab}{the y-label. Defaults to "Score".}

\item{xlim}{the x-range. Selected automatically unless specified.}

\item{ylim}{the y-range. Selected automatically unless specified.}
}
\value{
a plot of scores by leadtime with one curve for each forecast candidate.
}
\description{
Plot comparable variogram scores
}
