% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_contingency_table.R
\name{print_contingency_table}
\alias{print_contingency_table}
\title{Contingency table}
\usage{
print_contingency_table(ctab_list)
}
\arguments{
\item{ctab_list}{List of contingency tables}
}
\value{
Prints contingency table(s) for all forecast candidates.
}
\description{
Contingency table
}
