% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printContingencyTable.R
\name{printContingencyTable}
\alias{printContingencyTable}
\title{Contingency table}
\usage{
printContingencyTable(ctab_list)
}
\arguments{
\item{ctab_list}{List of contingency tables}
}
\value{
Prints contingency table(s) for all forecast candidates.
}
\description{
Contingency table
}
