% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_table_list.R
\name{probability_table_list}
\alias{probability_table_list}
\title{Converts a list of event detections tables into a list of probability tables. Intermediate function.}
\usage{
probability_table_list(detect_table_list)
}
\arguments{
\item{detect_table_list}{A list of detection tables.}
}
\value{
A list of probability tables.
}
\description{
Converts a list of event detections tables into a list of probability tables. Intermediate function.
}
