% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilityTableList.R
\name{probabilityTableList}
\alias{probabilityTableList}
\title{Converts a list of event detections tables into a list of probability tables. Intermediate function.}
\usage{
probabilityTableList(detect_table_list)
}
\arguments{
\item{detect_table_list}{List of detection tables}
}
\value{
A number.
}
\description{
Converts a list of event detections tables into a list of probability tables. Intermediate function.
}
