% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantilePlot.R
\name{quantilePlot}
\alias{quantilePlot}
\title{Produce fanchart / quantile plot}
\usage{
quantilePlot(
  m,
  main = "Quantile Plot",
  xlab = "Time",
  ylab = "Quantity",
  xlim = 0,
  ylim = 0,
  dt = NULL,
  grid = T,
  col = c("light blue", "dark blue"),
  dt.type = "normal"
)
}
\arguments{
\item{m}{To be written in 2024.}

\item{main}{To be written in 2024.}

\item{xlab}{To be written in 2024.}

\item{ylab}{To be written in 2024.}

\item{xlim}{To be written in 2024.}

\item{ylim}{To be written in 2024.}

\item{dt}{To be written in 2024.}

\item{grid}{To be written in 2024.}

\item{col}{To be written in 2024.}

\item{dt.type}{To be written in 2024.}
}
\value{
Does not return anything yet, but displays a quantile plot.
}
\description{
Produce fanchart / quantile plot
}
