% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_histogram_list.R
\name{rank_histogram_list}
\alias{rank_histogram_list}
\title{Rank histograms of a list of forecast candidates.}
\usage{
rank_histogram_list(
  data,
  nbins = 0,
  seed = 0,
  xlab = "Transformed ranks",
  ylab = "Frequency"
)
}
\arguments{
\item{data}{A number.}

\item{nbins}{The desired number of bins. Defaults to m+1 bins (code 0).}

\item{seed}{Optional seed. Defaults to 0 (random seed every time).}

\item{xlab}{x-axis label. Defaults to "Transformed ranks".}

\item{ylab}{y-axis label. Defaults to "Frequency".}

\item{main}{Title of the histogram.}
}
\value{
Rank histogram for every forecast candidate in the data structure.
}
\description{
Rank histograms of a list of forecast candidates.
}
