% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankHistogram.R
\name{rankHistogram}
\alias{rankHistogram}
\title{Rank histogram of a series of ensemble forecasts w.r.t a series of observations.}
\usage{
rankHistogram(f, y, nbins = 0, seed = 0)
}
\arguments{
\item{f}{Vector representing a series of ensemble forecasts.}

\item{y}{The corresponding observation series.}

\item{nbins}{The desired number of bins. Defaults to m+1 bins (code 0)}

\item{seed}{Optional seed. Defaults to 0 (random seed every time)}
}
\value{
Rank histogram
}
\description{
Rank histogram of a series of ensemble forecasts w.r.t a series of observations.
}
