% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankHistogramList.R
\name{rankHistogramList}
\alias{rankHistogramList}
\title{Rank histograms of a list of forecast candidates.}
\usage{
rankHistogramList(
  data,
  nbins = 0,
  seed = 0,
  xlab = "Transformed ranks",
  ylab = "Frequency"
)
}
\arguments{
\item{data}{A number. More parameters to do.}
}
\value{
A number.
}
\description{
Rank histograms of a list of forecast candidates.
}
