% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{Compute the data points for a reliability diagram}
\usage{
reliability(pred_prob, binary_obs, bins = "default")
}
\arguments{
\item{pred_prob}{a vector of predicted probabilities.}

\item{binary_obs}{a vector of binary observations.}

\item{bins}{the vector of bins, given as the points of distinction between individual bins, including 0 and 1.}
}
\value{
a data frame with the points used for creating the reliability diagram.
}
\description{
Compute the data points for a reliability diagram
}
