% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_diagram_list.R
\name{reliability_diagram_list}
\alias{reliability_diagram_list}
\title{Reliability diagrams of a list of forecast candidates}
\usage{
reliability_diagram_list(
  detections,
  bins = "default",
  method = "classic",
  prob = F
)
}
\arguments{
\item{detections}{A list of detections, which can be either ensemble detection tables or probability tables}

\item{bins}{the vector of bins, given as the points of distinction between individual bins, including 0 and 1. The default binning is selected by setting this argument to "default".}

\item{method}{either classic approach ("classic") or CORP ("CORP"). Defaults to "classic".}

\item{prob}{Logical argument indiciating whether the first argument are probabilities or ensemble detections.}
}
\value{
hmm...
}
\description{
Reliability diagrams of a list of forecast candidates
}
