% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityDiagram.R
\name{reliabilityDiagram}
\alias{reliabilityDiagram}
\title{Produce a reliability diagram}
\usage{
reliabilityDiagram(
  pred_prob,
  binary_obs,
  bins = "default",
  main = "Reliability diagram",
  xlab = "Forecast probability",
  ylab = "Observed relative frequency"
)
}
\arguments{
\item{pred_prob}{a vector of predicted probabilities.}

\item{binary_obs}{a vector of binary observations.}

\item{bins}{the vector of bins, given as the points of distinction between individual bins, including 0 and 1.}

\item{main}{the title of the plot. Defaults to "Reliability diagram".}

\item{xlab}{the x-label. Defaults to "Forecast probability".}

\item{ylab}{the y-label. Defaults to "Observed relative frequency".}
}
\value{
a reliability diagram produced from the input data.
}
\description{
Produce a reliability diagram
}
