% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_curve.R
\name{roc_curve}
\alias{roc_curve}
\title{Produce a ROC curve for}
\usage{
roc_curve(detect_table_prob, main = "")
}
\arguments{
\item{detect_table_prob}{detection table with event probabilities}

\item{main}{title of ROC plot}
}
\value{
Contingency table for all forecast candidates
}
\description{
Produce a ROC curve for
}
