% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurve.R
\name{rocCurve}
\alias{rocCurve}
\title{ROC Curve}
\usage{
rocCurve(detect_table_sum, main = "")
}
\arguments{
\item{detect_table_sum}{detection table with probabilities}

\item{main}{title of ROC plot}
}
\value{
Contingency table for all forecast candidates
}
\description{
ROC Curve
}
