% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurveList.R
\name{rocCurveList}
\alias{rocCurveList}
\title{ROC Curve}
\usage{
rocCurveList(detect_table_list, main = "")
}
\arguments{
\item{main}{title of ROC plot}

\item{detect_table}{List of detection tables}

\item{threshold}{Required number of positive ensemble members}
}
\value{
Contingency table for all forecast candidates
}
\description{
ROC Curve
}
