% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_fc_obs_data}
\alias{sample_fc_obs_data}
\title{Sample data included in package}
\format{
list with two entries, \code{observations} and \code{forecasts}. 
\code{forecasts} is a list of \code{data.frame}s containing forecast data.
}
\usage{
sample_fc_obs_data
}
\description{
Sample data included in package
}
\keyword{data}
