% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghettiPlot.R
\name{spaghettiPlot}
\alias{spaghettiPlot}
\title{Produce spaghetii plot}
\usage{
spaghettiPlot(
  m,
  main = "Spaghetti Plot",
  xlab = "Time",
  ylab = "Quantity",
  xlim = 0,
  ylim = 0,
  dt = NULL,
  grid = T,
  col = "gray",
<<<<<<< Updated upstream
  dt.type = "normal"
=======
  ...
>>>>>>> Stashed changes
)
}
\arguments{
\item{m}{To be written in 2024.}

\item{main}{To be written in 2024.}

\item{xlab}{To be written in 2024.}

\item{ylab}{To be written in 2024.}

\item{xlim}{To be written in 2024.}

\item{ylim}{To be written in 2024.}

\item{dt}{To be written in 2024.}

\item{grid}{To be written in 2024.}

\item{col}{To be written in 2024.}

\item{dt.type}{To be written in 2024.}
}
\value{
Does not return anything yet, but displays a quantile plot.
}
\description{
Produce spaghetii plot
}
