% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Print summary statistics}
\usage{
summary_stats(data)
}
\arguments{
\item{data}{A list with two elements. The first element contains the forecast data, and the second contains observations.}
}
\value{
Does not return anything, but simply prints the summary statistics on the screen.
}
\description{
Print summary statistics
}
