% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{summaryStats}
\alias{summaryStats}
\title{Print summary statistics}
\usage{
summaryStats(data)
}
\arguments{
\item{data}{A list with two elements. The first element contains forecast data as a list, and the second contains observations.}
}
\value{
Does not return anything, but simply prints the summary statistics on the screen.
}
\description{
Print summary statistics
}
