% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{vars}
\alias{vars}
\title{VarS of a multivariate probabilistic forecast w.r.t a vector of observations.}
\usage{
vars(f, y, p)
}
\arguments{
\item{f}{Vector representing a probabilistic forecast.}

\item{y}{The corresponding observation.}

\item{p}{The order. Defaults to 0.5.}
}
\value{
Variogram score (VarS) of f and y.
}
\description{
VarS of a multivariate probabilistic forecast w.r.t a vector of observations.
}
