% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_score_list.R
\name{brier_score_list}
\alias{brier_score_list}
\title{Brier scores for a list of event detections}
\usage{
brier_score_list(detections, prob = F)
}
\arguments{
\item{detections}{A list of detections, which can be either ensemble detection tables or probability tables}

\item{prob}{Logical argument indiciating whether the first argument are probabilities or ensemble detections.}
}
\value{
A vector of Brier scores, one for each forecast model.
}
\description{
Brier scores for a list of event detections
}
