% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_range.R
\name{check_for_range}
\alias{check_for_range}
\title{Check for range-hits within a univariate time series subset}
\usage{
check_for_range(data, range)
}
\arguments{
\item{data}{A vector of numerical values}

\item{range}{Range/interval of forecasted variable required for having an event}
}
\value{
Boolean, 1 = event happened, 0 = event did not happen
}
\description{
Check for range-hits within a univariate time series subset
}
