% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventDetect.R
\name{eventDetect}
\alias{eventDetect}
\title{Function for detecting all events in a multivariate time series}
\usage{
eventDetect(data, change = NA, range = NA, window)
}
\arguments{
\item{data}{Data object returned by the loadData function}

\item{change}{Flat signed change in forecasted variable}

\item{range}{?}

\item{window}{Time window length within which the change occurs}
}
\value{
Returns an event detection table
}
\description{
Function for detecting all events in a multivariate time series
}
