% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventDetectionTable.R
\name{eventDetectionTable}
\alias{eventDetectionTable}
\title{Function for constructinga list of event detection tables.}
\usage{
eventDetectionTable(data, change, window)
}
\arguments{
\item{data}{Data object returned by the loadData function}

\item{change}{Flat signed change in forecasted variable}

\item{window}{Time window length within which the change occurs}
}
\value{
Returns an event detection table
}
\description{
Function for constructinga list of event detection tables.
}
