% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastEvaluationReport.R
\name{forecastEvaluationReport}
\alias{forecastEvaluationReport}
\title{DO NOT USE THIS YET}
\usage{
forecastEvaluationReport(data, dest = "./", options = NULL)
}
\arguments{
\item{data}{Data set in the required format.}

\item{dest}{Destination for the compiled HTML report.}
}
\value{
Currently nothing.
}
\description{
DO NOT USE THIS YET
}
