% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLeadTime.R
\name{getLeadTime}
\alias{getLeadTime}
\title{Add together two numbers}
\usage{
getLeadTime(forecast)
}
\arguments{
\item{forecast}{A forecast table or row containing at least the columns named TimeStamp and BaseTime, respectively.}
}
\value{
Lead times of the input forecasts.
}
\description{
Add together two numbers
}
