% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_evaluation_subset.R
\name{make_evaluation_subset}
\alias{make_evaluation_subset}
\title{Refine data to evaluatable dataset}
\usage{
make_evaluation_subset(data)
}
\arguments{
\item{data}{A forecast-observation data structure.}
}
\value{
A forecast-observation data structure with only intersecting timestamps preserved.
}
\description{
Refine data to evaluatable dataset
}
