% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFc.R
\name{plotFc}
\alias{plotFc}
\title{Plot forecast}
\usage{
plotFc(
  data,
  f_id,
  style = "fanchart",
  xmax = 100,
  main = "Forecast",
  xlab = "Time",
  ylab = "Quantity",
  xlim = 0,
  ylim = 0,
  grid = T,
  col = NULL,
  dt.type = "normal"
)
}
\arguments{
\item{data}{Data object returned by the loadData function}

\item{f_id}{forecast candidate}

\item{style}{fanchart or spaghetti, defaults to fanchart}

\item{xmax}{Maximum number of observations plotted}

\item{main}{To be written in 2024.}

\item{xlab}{To be written in 2024.}

\item{ylab}{To be written in 2024.}

\item{grid}{To be written in 2024.}

\item{col}{To be written in 2024.}

\item{dt.type}{To be written in 2024.}
}
\value{
Does not return anything yet, but displays a quantile plot.
}
\description{
Plot forecast
}
