% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityDiagramList.R
\name{reliabilityDiagramList}
\alias{reliabilityDiagramList}
\title{Reliability diagrams of a list of forecast candidates}
\usage{
reliabilityDiagramList(
  detections,
  bins = "default",
  method = "classic",
  prob = F
)
}
\arguments{
\item{detections}{A list of detections, which can be either ensemble detection tables or probability tables}

\item{prob}{Logical argument indiciating whether the first argument are probabilities or ensemble detections.}
}
\value{
hmm...
}
\description{
Reliability diagrams of a list of forecast candidates
}
