% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_curve_list.R
\name{roc_curve_list}
\alias{roc_curve_list}
\title{Produce ROC curves for every forecast candidate in an evaluation set}
\usage{
roc_curve_list(
  detect_table_list,
  main = NA,
  col.curve = "black",
  col.auc = "light gray",
  xlab = "False positive rate",
  ylab = "True positive rate"
)
}
\arguments{
\item{detect_table_list}{List of detection tables}

\item{main}{title of ROC plot}
}
\value{
Contingency table for all forecast candidates
}
\description{
Produce ROC curves for every forecast candidate in an evaluation set
}
