% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurve.R
\name{rocCurve}
\alias{rocCurve}
\title{ROC Curve}
\usage{
rocCurve(detect_table, threshold = 5, main = "")
}
\arguments{
\item{threshold}{Required number of positive ensemble members}

\item{main}{title of ROC plot}

\item{detect_table_list}{List of detection tables}
}
\value{
Contingency table for all forecast candidates
}
\description{
ROC Curve
}
