% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghettiPlot.R
\name{spaghettiPlot}
\alias{spaghettiPlot}
\title{Produce spaghetii plot}
\usage{
spaghettiPlot(
  m,
  x = NULL,
  main = "Spaghetti Plot",
  xlab = "Time",
  ylab = "Quantity",
  grid = T,
  col = "gray",
  ...
)
}
\arguments{
\item{m}{Ensemble forecast data}

\item{x}{Optional time stamps corresponding to rows of \code{m}.}

\item{main}{a main title for the plot, see also \code{?plot}.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{grid}{Add a grid to the plot background?}

\item{col}{Ends of the colour range for fan plot.}

\item{...}{Additional arguments passed to \code{plot()}}
}
\value{
Does not return anything, but displays a quantile plot.
}
\description{
Produce spaghetii plot
}
