% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram_score.R
\name{variogram_score}
\alias{variogram_score}
\title{VarS of a multivariate probabilistic forecast w.r.t a vector of observations.}
\usage{
variogram_score(
  f,
  y,
  p,
  weights = function(i, j) {
     1/abs(i - j)
 }
)
}
\arguments{
\item{f}{A matrix representing a multivariate probabilistic forecast. The row size is equal to the number of ensemble members, and the column size represents the dimension.}

\item{y}{The corresponding observation.}

\item{p}{The order. Defaults to 0.5.}

\item{weights}{Function for the weights of the (i,j)-pairs. Defaults to 1/|i-j|.}
}
\value{
Variogram score (VarS) of f and y.
}
\description{
VarS of a multivariate probabilistic forecast w.r.t a vector of observations.
}
